refactor add/move reflect template prompt to sdl
