+++ packages/server/graphql/public/types/MoveReflectTemplatePromptPayload.ts
@@ -1,0 +1,15 @@
+import {MoveReflectTemplatePromptPayloadResolvers} from '../resolverTypes'
+
+export type MoveReflectTemplatePromptPayloadSource =
+  | {
+      promptId: string
+    }
+  | {error: {message: string}}
+
+const MoveReflectTemplatePromptPayload: MoveReflectTemplatePromptPayloadResolvers = {
+  prompt: (source, _args, {dataLoader}) => {
+    return 'promptId' in source ? dataLoader.get('reflectPrompts').load(source.promptId) : null
+  }
+}
+
+export default MoveReflectTemplatePromptPayload
